feat: only consider known tokens
